Modify Popup system to use IDs
